Verify events' byte length and clear previous data
Compares the number of bytes reported in the header associated with each
event block, and the actual number of bytes in each event block, and
only processes those events whose values are in agreement.

Also clears the subprocess output files "wtcoeffs" and
"wtcoeffs.icdf24_?" so that inverted data from a previous event is not
read in and associated with the current event. This is not as important
now that the byte-length check is performed, but the code will now fail
with a helpful prinout instead of incorrectly proceeding.