Refactor linear interpolation method
This will allow a single GPS instance to be input in place of a list of
inputs, as is necessary for the next version of automaid

NB, I did not change the outcome of the algorithm at all; I just
reorganized it to accept more inputs.

Also adds plain-English description of the outcome of interpolation via
the attached interpolation dictionary value in "description"; prints
this alongside any failures in *gps_linear_interpolation.txt.