Allow single GPS for surface-drift interpolation
If one GPS point is passed to gps.linear_interpolation then no
interpolation is performed and that input is simply returned.

I argue this is valid because we still will always pass two GPS points
to the important piece: drift in the mixed layer.  See comment in code
for example.