Return known, not requested, date if no interp.
Previously if input points were near in LOCATION, then interpolation was
not performed and the known location was returned.  However, the
proximity in TIME of the known points to the interpolation (requested)
date was not considered.

This could be problematic if the input GPS list included points very
near one another in location (and thus interpolation was not
performed), but their times were very far from the requested
(interpolation) time.

It is my opinion that is is better in the latter case to return the
known location AND TIME, as opposed to assigning the known location to
any random time (potentially infinitely far in the past or future).