Convert SAC metadata floats to numpy.float32()
I realized the output loc.txt STLA and STLO locations did NOT match
those written to the SAC headers.  Fixed it with conversion of those
native floats to numpy.float32.

I ran some tests and found that:
    self.sacmeta["stla"] = self.station_loc.latitude;
    self.sacmeta["stlo"] = self.station_loc.longitude;
vs.
    self.sacmeta["stla"] = np.float32(self.station_loc.latitude);
    self.sacmeta["stlo"] = np.float32(self.station_loc.longitude);

does not change the SAC files, i.e., somewhere in obspy they convert
this to float32(? I haven't identified that specific line...), but it
DOES change the loc.txt file, because "{.6f}".format(...) prints
something different depending on if the input is a native float or a
numpy.float32.