Update scale factor to convert counts to pascals
From Seb 24-May-2019:
"The factor 5 is equal to the maximum voltage amplitude at the
input of the convertor that is 2.5V multiplied by 2 because
there is 2 channels (differential input). But the electronic
gain of 25 dB probably does not take account of the
differential input, then I should add a factor 2 in the
equation:
10**((-201.+25.)/20.) * 2 * 2**28/5. * 1000000 = 170176"

sensitivity of the hydrophone: -201 dB re V/uPa
electronic gain: 25 dB
analog to digital conversion: 2**28/5. count/V
count/uPa to count/Pa conversion: 1000000