Define elev(STEL), azimuth(CMPAZ), SACdip(CMPINC)
Set all to 0, based on what IRIS serves me up for ?DH components.

This looks like a major commit but it's not; there is so much deleted
text because I reworked how I defined both the mseed2sac_metadata and
the automaid_metadata text file to use str joins and regexp
substitutions (so that I didn't have to define all four separately).