Combine .LOG and .MER to generate unique GPS pairs
*GENERALLY* MERMAID GPS fixes are written as (nonunique) pairs whose
timing and locations differ only on the order of seconds and meters.
One GPS fix is written to the .LOG and one to the .MER.

Seb told me ideally we take the timing info from the .MER and the
position info from the .LOG.

The object `GPS_MERGED` is like a standard `GPS` object except that it
may use timing/location data from both the .LOG and .MER for each GPS
pair.  There are generally 1/2 the `GPS_MERGED` instances as `GPS`
instances.

The method `merge_gps_list` pairs up the nonunique individual-file GPS
fixes. Examples of how various GPS lists are paired may be found in
'example_merge_gps_list.py'