Leave/reach surface LAYER in gps_interpolation.txt
Previously was setting:
 leg_descent = dive.descent_leave_surface_loc
 leg_ascent = dive.ascent_reach_surface_loc

not
 leg_descent = dive.descent_leave_surface_layer_loc
 leg_ascent = dive.ascent_reach_surface_layer_loc

so it was only including the time/drift up to leaving from the surface,
not the surface LAYER (50 m).  So the velocities should all be the same,
just the timing should be greater with this new commit, where I also
rephrase to use:

 leg_descent = dive.descent_last_loc_before_event
 leg_ascent = dive.ascent_first_loc_after_event

which are now set to either be the surface or surface-layer locations
where MERMAID dove/ascended.