Patch for P0006 and P0007
Here I noticed that seemingly redundant REQ data were being transmitted
in P0006 and P0007 .MER files.  Upon further inspection I realized that
in one case the data were in fact redundant, while in another they
differed slightly.

The former case required that I skip that event so that final
event-printout tallies were correct.  The latter case required a rename
of the output .sac and .mer files due to the seconds-precision of the
export name and the fractional-seconds differences between the two REQ
signals.  For reference:

    Same file name, redundant data (remove second event from list):
	'20180728T225619.07_5B7739F0.MER.REQ.WLT5'
    Same export file name  (.sac, .mseed), different data (do not remove from list!):
	'20180728T225619.06_5B773AE6.MER.REQ.WLT5'
    Rename second occurrence of redundant EXPORT file name to:
	'20180728T225619.06_5B773AE6.MER1.REQ.WLT5'

Note that the latter case only generated a single output line in the
GeoCSV because that only only shows seconds precision.  Note as well
that the latter case resulted in the same location in loc.txt, precise
there to six decimal places.

Known issues:
*tool_invert_mer.py no longer works (possibly broken before this commit)