Define `pressure_mbar` attr for dives.Complete_Dive
Ulimately will be passed to GeoCSV and written there.

At the moment this is not very efficient (none of it is) because the
same line --

`utils.find_timestamped_values("P\s*(\+?\-?\d+)mbar",self.log_content)`

-- is called for a Dive (via `generate_dive_plotly` method) and Complete
Dive instance, so we parse the mbar readings from the LOG files twice
(maybe cleanup at some point...but this code is pretty inefficient as it
is and files are reread all over the place...).