Remove lines from LOG file with bad epoch times
.LOG files:

    25_643FB6EF.LOG
    25_8929B84B.LOG
    25_8929BED9.LOG

experienced clock errors that resulted in negative Unix epoch times (integer
seconds preceding ":" in each line).  The fix implemented here checks the
purported time that each .LOG line was written, and removes it if it precedes the
start time of the dive, as indicated by hex in the dive name itself.

Had to separate `split_log_lines` into `get_log_delimiter` and `split_log_lines`
to separately parse the delimiter from the LOG file in case I had to reconstruct
it later by concatenated a winnowed .LOG-content list of strings.