Add "Algorithm(thermocline)" rows to GeoCSV
This is the interpolated time and lat/lon of crossing the thermocline
(currently 50 m depth) on descent and ascent. Note the pressure in mbar
of that passage is not actually measured, its just set to the
"mixed layer depth" (surface layer) in meters and multiplied by 100 to
convert to mbar.